Fix: correção de estrutura de classes, adição de códigos http e remoção de arquivos não utilizados
